fairchain-ai/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.9.2
python-dotenv>=1.0.0
httpx>=0.28.1
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.2
scipy>=1.13.0
shap>=0.46.0
google-genai>=1.7.0
google-generativeai>=0.8.0